backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-api-python-client==2.86.0
vaderSentiment==3.3.2
transformers==4.21.0
torch==1.9.0
redis==4.3.4
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
textblob
google-api-python-client==2.86.0
vaderSentiment==3.3.2
transformers==4.21.0
torch==1.9.0
redis==4.3.4
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
textblob
transformers==4.21.0
torch==1.9.0
redis==4.3.4
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
